engine: publicly exposed the task API
Expose it as an internal API. The reason why we need to expose
this internal C++ API is that strings in Android (and in Java
of course) are not UTF-8 but modified UTF-8.

This poses issues such as the one described at #1877.

I have figured out that the best way to deal with this issue
is to let JNI code talk directly with MK's C++ code such that
the notion that strings are zero terminate could be removed
from the whole codebase. This means, in turn, that it will
be potentially possible to create byte arrays with confidence
and use such byte arrays rather than using strings.

More tricks are possible, such as forcing a specific JSON
serialisation that only emits ASCII.

Most of that is already doable using the C string API except
that that seems brittle because strings are zero terminated
in C and we cannot guarantee the strings returned by MK won't
contain zeros in the future. That is, as unlikely as this
could be, using a better API, i.e. the C++ string API, where
that _cannot_ happen seems a better solution.
